+++ src/main/java/de/uni_jena/cs/fusion/abecto/processor/implementation/SparqlConstructProcessor.java
@@ -15,13 +15,16 @@
  */
 package de.uni_jena.cs.fusion.abecto.processor.implementation;
 
+import org.apache.jena.ontology.OntModel;
 import org.apache.jena.query.Query;
 import org.apache.jena.query.QueryExecution;
 import org.apache.jena.query.QueryExecutionFactory;
 import org.apache.jena.query.QueryFactory;
+import org.apache.jena.rdf.model.Model;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
+import de.uni_jena.cs.fusion.abecto.model.Models;
 import de.uni_jena.cs.fusion.abecto.parameter_model.ParameterModel;
 import de.uni_jena.cs.fusion.abecto.processor.AbstractTransformationProcessor;
 
@@ -32,15 +35,29 @@ public class SparqlConstructProcessor extends AbstractTransformationProcessor<Sp
 		// prepare query
 		Query query = QueryFactory.create(this.getParameters().query);
 
-		// prepare execution
-		QueryExecution queryExecution = QueryExecutionFactory.create(query, this.inputModelUnion);
+		OntModel inputAndResultModelUnion = Models.getEmptyOntModel();
+		inputAndResultModelUnion.addSubModel(this.inputModelUnion);
+		inputAndResultModelUnion.addSubModel(this.getResultModel());
 
-		// execute write into result model
-		queryExecution.execConstruct(this.getResultModel());
+		for (int iteration = 1; iteration <= this.getParameters().maxIterations; iteration++) {
+			// prepare execution
+			QueryExecution queryExecution = QueryExecutionFactory.create(query, inputAndResultModelUnion);
+
+			// execute and write into intermediate result model
+			Model intermediateResultModel = queryExecution.execConstruct(Models.getEmptyOntModel());
+
+			// add new statements (if any) to result model, otherwise break
+			if (!this.getResultModel().containsAll(intermediateResultModel)) {
+				this.getResultModel().add(intermediateResultModel);
+			} else {
+				break;
+			}
+		}
 	}
 
 	@JsonSerialize
 	public static class Parameter implements ParameterModel {
 		public String query;
+		public int maxIterations = 1;
 	}
 }